data_collector/requirements.txt
aiohttp==3.9.5
shodan==1.31.0
censys==2.2.14
pydantic==2.6.4
pydantic-settings==2.2.1
apscheduler==3.10.4
pandas==2.2.2
numpy==1.26.4
pyyaml==6.0.1
